circular component dependencies - working (yay)
