+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/cell.component.html
@@ -1,0 +1,41 @@
+<ng-container *ngIf="formGroup && mode && cell && cell.attribute">
+  <ng-container [ngSwitch]="cell.attribute.type">
+
+    <ng-container *ngSwitchCase="'Date'">
+      <app-date-picker-control [formGroup]="formGroup" [mode]="mode" [attribute]="cell.attribute"></app-date-picker-control>
+    </ng-container>
+
+    <ng-container *ngSwitchCase="['Text', 'Html'].includes(cell.attribute.type) ? cell.attribute.type : ''">
+      <ng-container [ngSwitch]="cell.component">
+        <ng-container *ngSwitchCase="'TextArea'">
+          <app-text-area-control [cell]="cell" [formGroup]="formGroup" [mode]="mode" [attribute]="cell.attribute"></app-text-area-control>
+        </ng-container>
+
+        <ng-container *ngSwitchDefault>
+          <lib-new-text-field-control [formGroup]="formGroup" [mode]="mode" [name]="cell.attribute.name"></lib-new-text-field-control>
+        </ng-container>
+      </ng-container>
+    </ng-container>
+
+    <ng-container *ngSwitchCase="'OneToMany'">
+      <lib-new-one-to-many-control [mode]="mode" [cell]="cell" [formGroup]="formGroup"></lib-new-one-to-many-control>
+    </ng-container>
+
+    <ng-container *ngSwitchCase="'ManyToOne'">
+      <ng-container [ngSwitch]="cell.component">
+        <ng-container *ngSwitchCase="'Select'">
+          <app-select-control [formGroup]="formGroup" [mode]="mode" [attribute]="cell.attribute"></app-select-control>
+        </ng-container>
+
+        <ng-container *ngSwitchDefault>
+          <app-many-to-one-control [formGroup]="formGroup" [mode]="mode" [attribute]="cell.attribute"></app-many-to-one-control>
+        </ng-container>
+      </ng-container>
+    </ng-container>
+
+    <ng-container *ngSwitchDefault>
+      ERROR: Missing or unknown attribute type of: "{{cell.attribute.type}}"
+    </ng-container>
+
+  </ng-container>
+</ng-container>
