+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/controls/label/label.component.html
