+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/controls/new-one-to-one-control/new-one-to-one-control.component.html
