+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/controls/new-text-field-control/new-text-field-control.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {CellAttribute} from '../../../../../../meta/page/meta-page-service/meta-page.service';
+import {CellAttribute} from '../../../../../meta/page/meta-page-service/meta-page.service';
 import {FormGroup} from '@angular/forms';
 
 @Component({
