+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/new-one-to-many-control.component.html
@@ -1,12 +1,12 @@
 <ng-container *ngIf="cell && cell.template">
   <div class="border-top">
     <ng-container *ngIf="cell && cell.attribute && cell.attribute.relationshipTarget">
-      <!--<lib-layout
+      <lib-layout
         [mode]="mode"
         [template]="cell.template"
         [formGroup]="formGroup"
         [relationshipProperty]="'members'">
-      </lib-layout>-->
+      </lib-layout>
     </ng-container>
   </div>
 </ng-container>
