+++ angular-workspace/projects/ngx-perfect-stack/src/lib/data/controller/layout/new-table-layout.component.html
